fix: quantity type changed to Integer

+++ domain/src/main/java/org/capitalmind/dto/request/StockRequest.java
@@ -1,4 +1,4 @@
 package org.capitalmind.dto.request;
 
-public record StockRequest(String name, String description, Double quantity, String userId) {
+public record StockRequest(String name, String description, Integer quantity, String userId) {
 }
